Se seccionarion las clases y los metodos
