+++ ene-jun-2020/JOSE EDUARDO ONOFRE SUAREZ/ProyectoFinalYGODeckMaker/ProyectoFinalYGO/Card.py
@@ -1,0 +1,14 @@
+class Card:
+    def __init__(self, id, name, types, desc, atk, defs,level, race, attribute, price):
+
+        self.Id = id
+        self.Name = name
+        self.Type = types
+        self.Desc = desc
+        self.Atk = atk
+        self.Def = defs
+        self.Level = level
+        self.Race = race
+        self.Attribute = attribute
+        self.Price = price
+
