+++ ene-jun-2020/JOSE EDUARDO ONOFRE SUAREZ/ProyectoFinalYGODeckMaker/ProyectoFinalYGO/Collection.py
@@ -1,0 +1,11 @@
+
+
+
+def ShowMyDecks():
+    pass
+
+
+
+
+def countCards():
+    contador=0
